Add filesystem manipulation (file parsing, dir listing)
According to the rfc and the item types, this commit add
a way to handle the request made by the client manipulating
the resource in function of its type (dir/file).
An interesting feature is the ability to parse the line and
build a gopher_item based on the ext of the file.

Signed-off-by: fpantano <fpantano@redhat.com>