Improve Makefile targets and start working on main().
Start working on main parameters to create a minimal
working cli with the following parameters:
 1. host
 2. port
 3. srv path (still WIP)
The second aspect of this commit is represented by the
improvement of the Makefile to add several parameters to
be considered at compile time (e.g. VERSION and CFLAGS),
as well as a little helper to easily find the MK targets.

Signed-off-by: fpantano <fpantano@redhat.com>